requirements-adk.txt
# Google ADK and core dependencies
google-adk>=1.7.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Messaging and caching
kafka-python>=2.0.2
redis>=4.5.0

# HTTP and web
httpx>=0.24.0
requests>=2.31.0

# Data parsing
feedparser>=6.0.10
beautifulsoup4>=4.12.0

# Geospatial analysis
osmnx>=1.6.0
networkx>=3.1
geopandas>=0.14.0
shapely>=2.0.0
folium>=0.14.0

# Configuration
python-dotenv>=1.0.0 